+++ react/components/UserWidget.tsx
@@ -24,6 +24,9 @@ import SET_CURRENT_ORGANIZATION from '../graphql/setCurrentOrganization.graphql'
 import STOP_IMPERSONATION from '../graphql/impersonateUser.graphql'
 import { B2B_CHECKOUT_SESSION_KEY } from '../utils/constants'
 import '../css/user-widget.css'
+import { sendStopImpersonateMetric } from '../utils/metrics/impersonate'
+import type { ChangeTeamParams } from '../utils/metrics/changeTeam'
+import { sendChangeTeamMetric } from '../utils/metrics/changeTeam'
 
 const CSS_HANDLES = [
   'userWidgetContainer',
@@ -216,6 +219,13 @@ const UserWidget: VtexFunctionComponent<UserWidgetProps> = ({
           sessionStorage.removeItem(B2B_CHECKOUT_SESSION_KEY)
         }
 
+        const metricParams = {
+          sessionResponse,
+          email: userWidgetData?.checkImpersonation?.email,
+          ...organizationsState,
+        }
+
+        sendStopImpersonateMetric(metricParams)
         window.location.reload()
       })
       .catch(error => {
@@ -254,6 +264,13 @@ const UserWidget: VtexFunctionComponent<UserWidgetProps> = ({
           costId: organizationsState.currentCostCenter,
         },
       })
+
+      const metricParams: ChangeTeamParams = {
+        sessionResponse,
+        ...organizationsState,
+      }
+
+      sendChangeTeamMetric(metricParams)
     } catch (error) {
       setErrorOrganization(true)
     } finally {